refactor - app name
mudança de endereço do app pro github da organizacao + pastas melhor organizadas
