+++ android/app/src/main/kotlin/com/github/proanonimos/plataforma_profissional_anonima/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.plataforma_profissional_anonima
+package com.github.proanonimos.plataforma_profissional_anonima
 
 import io.flutter.embedding.android.FlutterActivity
